Add integration tests for json ingest
