+++ .github/workflows/build.yml
@@ -47,4 +47,4 @@ jobs:
         with:
           node-version: ${{ matrix.node }}
       - name: Build with Maven
-        run: mvn -B package
+        run: mvn -B verify
